Modify(changhoon/signup) : #28
	1. 폴더명 변경 (signup -> sign-up)
	2. 파일명 변경 (signUp.ctrl -> sign-up.ctrl)
	3. Error 모듈 적용
