+++ app/app.js
@@ -7,7 +7,7 @@ dotenv.config();
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
-const root = require('./src/apis/signup');
+const root = require('./src/apis/sign-up');
 const review = require('./src/apis/review');
 
 app.use('/api', root);
