+++ app/src/apis/sign-up/index.js
@@ -3,9 +3,9 @@
 const express = require('express');
 
 const router = express.Router();
-const ctrl = require('./signUp.ctrl');
+const ctrl = require('./sign-up.ctrl');
 
 // 회원가입 API
-router.post('/signUp', ctrl.process.signUp);
+router.post('/sign-up', ctrl.process.signUp);
 
 module.exports = router;
