+++ app/src/models/services/Student/Student.js
@@ -4,6 +4,7 @@ const bcrypt = require('bcrypt');
 
 const saltRounds = 10;
 const StudentStorage = require('./StudentStorage');
+const Error = require('../../utils/Error');
 
 class Student {
   constructor(body) {
@@ -53,14 +54,13 @@ class Student {
       }
       return {
         saveable: false,
-        msg: '서버 에러입니다. 서버개발자에게 문의하세요',
+        msg: '서버 에러입니다. 서버개발자에게 문의하세요.',
       };
     } catch (err) {
-      return {
-        saveable: false,
-        msg: '알 수 없는 오류입니다. 서버개발자에게 문의하세요',
-        error: err,
-      };
+      return Error.ctrl(
+        '알 수 없는 오류입니다. 서버개발자에게 문의하세요.',
+        err
+      );
     }
   }
 }
