+++ app/src/models/utils/Error.js
@@ -1,1 +1,13 @@
-// 에러파일
+'use strict';
+
+class Error {
+  static ctrl(msg, err) {
+    return {
+      isError: true,
+      errMsg: err,
+      clientMsg: msg,
+    };
+  }
+}
+
+module.exports = Error;
